Infrastructure: split TLuaInterpreter.cpp's UI files out separately (#7156)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Split TLuaInterpreter.cpp into a domain-specific UI functions file.
#### Motivation for adding to Mudlet
A 18k line C++ files makes a lot of tools cry.
#### Other info (issues closed, discussion etc)
This is the final PR in this series - TLuaInterpreter.cpp is now down to
8k lines from the initial 18k.

---------
